refactor: src/ 최상위 폴더 구조 재편 — api/, hooks/, lib/ 분리
- src/lib/api/ → src/api/ (서버 데이터 fetching)
- src/lib/hooks/ → src/hooks/ (React hooks)
- lib/에는 외부 라이브러리 래퍼만 유지 (dayjs, supabase-server, utils, menu-policy)
- day-bar.ts → _DayBar/ 컴포넌트 폴더 내부로 이동 (전용 유틸)
- hero-status.ts → HeroStatus/ 컴포넌트 폴더 내부로 이동 (전용 로직)
- getAnnouncements.ts → src/api/ 로 이동
- useHasMounted.ts, voterId.ts → src/hooks/ 로 이동
- 전체 import 경로 업데이트
