+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import {ErrorBoundary, SiteFooter, SiteHeader} from '@/components/@shared';
 import {MenuBoard} from '@/components/menu';
 import {HeroDate, HeroStatus, HomeSide} from '@/components/home';
-import getMenu from '@/lib/api/getMenu';
+import getMenu from '@/api/getMenu';
 import dayjs from '@/lib/dayjs';
 import {formatYYYYMMDD, getWeekDays} from '@/lib/utils';
